requirements.txt
albumentations==1.3.0
click==8.1.3
mlflow==2.4.2
numpy==1.23.5
Pillow==9.4.0
PyYAML==6.0
setuptools==65.6.3
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.65.0
